mediaTransformer/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "minio (>=7.2.15,<8.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic (>=2.11.4,<3.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "transformers (>=4.51.3,<5.0.0)",
    "num2words (>=0.5.14,<0.6.0)",
    "av (>=14.4.0,<15.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "redis (>=6.1.0,<7.0.0)",
]

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch"}
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.14"

